requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Admin Interface
sqladmin==0.21.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
# PostgreSQL drivers (choose one based on your environment)
psycopg2-binary==2.9.9  # For PostgreSQL (Linux/Mac)
# psycopg2==2.9.9  # Alternative for Windows if binary version fails
asyncpg==0.29.0  # Async PostgreSQL driver
aiosqlite==0.19.0  # Async SQLite driver
# sqlite3 is included in Python standard library

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.27.0
requests==2.31.0

# Environment & Configuration
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0  # Parallel test execution
pytest-html==4.1.1  # HTML test reports
pytest-benchmark==4.0.0  # Performance testing
httpx==0.27.0
factory-boy==3.3.0
responses==0.24.1
faker==20.1.0  # Generate fake test data
freezegun==1.2.2  # Mock datetime in tests
test-containers==3.7.1  # Integration testing with containers
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Optional: Background Tasks
arq==0.25.0
redis==5.0.1

# Process Management
honcho==1.1.0

# Optional: Caching
fastapi-cache2==0.2.1

# Optional: Documentation
mkdocs==1.5.3
mkdocs-material==9.5.0

# Monitoring
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
prometheus-api-client==0.5.2
pysqlite3==2.1.0